+++ src/divide/divide.hpp
@@ -1,6 +1,6 @@
 #ifndef DIVID_HPP
 #define DIVID_HPP 
 
-int divide(int a, int b);
+double divide(int a, int b);
 
 #endif /* ifndef DIVID_HPP */
